+++ firebase/functions/package.json
@@ -14,15 +14,15 @@
   },
   "main": "lib/Firebase/functions/src/index.js",
   "dependencies": {
-    "@google-cloud/firestore": "^4.4.0",
+    "@google-cloud/firestore": "^4.11.0",
     "csv-stringify": "^5.5.1",
-    "firebase-admin": "^9.2.0",
+    "firebase-admin": "9.7.0",
     "firebase-functions": "^3.6.1"
   },
   "devDependencies": {
+    "firebase-functions-test": "^0.2.0",
     "tslint": "^5.12.0",
-    "typescript": "^3.8.0",
-    "firebase-functions-test": "^0.2.0"
+    "typescript": "^3.8.0"
   },
   "private": true
 }
